requirements.txt
Quart==0.19.9
quart-cors==0.7.0
quart-openapi==1.7.2
Hypercorn==0.17.3
pyjwt
jsonschema==4.23.0
#grpc
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.3

requests==2.32.3
python-dotenv==1.0.1

black==24.10.0
aiofiles==24.1.0
aiohttp==3.11.12
quart-rate-limiter==0.11.0
openai==1.61.1
PyMuPDF==1.25.3
beautifulsoup4==4.13.3
inflect==7.5.0
nltk==3.9.1
libcst==1.6.0